Start over with new implementation.
Parallelising was indicating that the code had to be more modular. For
instance, we need to be able to loop over a range. Which was leading to the
calculation of the range, and hence the use of the modulo operation.
This is when I stumble upon http://stackoverflow.com/a/23045070/421846.